requirements.txt
# Oracle connectivity (pure Python thin mode - no Instant Client needed)
python-oracledb>=2.0.0

# Configuration & environment
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Graph algorithms for JOIN_PATH computation
networkx>=3.2.0

# String similarity for SIMILAR_TO edge inference
python-Levenshtein>=0.23.0

# Async support
anyio>=4.2.0

# LangChain / LangGraph agentic pipeline
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-anthropic>=0.3.0
openai>=1.50.0
anthropic>=0.40.0

# SQL parsing and dialect translation
sqlglot>=25.0.0

# Streamlit frontend
streamlit>=1.40.0
pandas>=2.0.0
plotly>=5.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
